x/orbit/types/messages_star_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/lubtd/orbit/testutil/sample"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateStar_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateStar
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateStar{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateStar{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateStar_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateStar
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateStar{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateStar{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteStar_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteStar
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteStar{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteStar{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
